\docType{package}
\name{AsthmaNHANES-package}
\alias{AsthmaNHANES}
\alias{AsthmaNHANES-package}
\title{Asthma Data Sets from National Health and Nutritional Examination Survey (NHANES)}
\description{
This package provides two datasets (metal and percent) that are related to the work of "Serum cadmium and lead, current wheeze, and lung function in a nationwide study of adults in the United States" (Journal of Allergy and Clinical Immunology: In Practice, DOI:10.1016/j.jai-p.2019.05.029).
}
\source{
Centers for Disease Control and Prevention (CDC). National Center for Health Statistics (NCHS). National Health and Nutrition Examination Survey Data. Hyattsville, MD: U.S. Department of Health and Human Services, Centers for Disease Control and Prevention, [2021][\url{https://www.cdc.gov/nchs/nhanes/}].
}
\author{
\itemize{
  \item Tao Sun
  \item Qiyao Qin
  \item Zihan Qian
  \item Yang Li
}}

